+++ Decisions.Equifax/PrequalificationOfOne/Request/PreQualificationOfOneRequest.cs
@@ -1,10 +1,9 @@
 ﻿using Decisions.Equifax.ConsumerCreditReport.Request;
 
-namespace Decisions.Equifax.PrequalificationOfOne.Request
+namespace Decisions.Equifax.PreQualificationOfOne.Request
 {
     public class PreQualificationOfOneRequest : ConsumerCreditReportRequest
-
     {
-    /*Since the request are the same, extending the Request class of Consumer Credit Report Request*/
+        /*Since the request are the same, extending the Request class of Consumer Credit Report Request*/
     }
 }
